requirements.txt
pandas
openpyxl
Flask
Flask-SQLAlchemy
flask_cors
gunicorn
python-dotenv
pyodbc
requests
urllib3